ajouter la sécurité de vérification pour modifier / supprimer (post + comment) + redirection après login/signup  sur la liste des posts + ajout defaut value pour photoURL

+++ backend/controllers/comment.js
@@ -45,6 +45,10 @@ exports.deleteComment = (req, res, next) => {
         return res.status(404).json({ error: "Commentaire non trouvé !" });
       }
 
+      if (req.auth.userId != comment.userId) {
+        return res.status(401).json({ error: "Suppression non autorisée !" });
+      }
+
       Comment.destroy({ where: { id: req.params.id } })
         .then((comment) =>
           res.status(200).json({ message: "Commentaire supprimé !" })
